+++ Sources/UI/PaymentMethods/SepaDataView.swift
@@ -114,11 +114,17 @@ public struct SepaDataEditorView: View {
                         CountryPicker(selectedCountry: $model.countryCode)
                     }
                 },
+                header: {
+                    Text(keyed: "Snabble.SEPA.helper")
+                        .font(.footnote)
+                        .foregroundColor(.secondaryLabel)
+                },
                 footer: {
                     Text(model.hintMessage)
                         .font(.footnote)
                         .foregroundColor(.red)
                 })
+            .textCase(nil)
             Section(
                 content: {
                     button